modified:   example-output-to-oled/patch.xodp 	new file:   example-set-address/patch.xodp 	modified:   example-test-configuration/patch.xodp 	modified:   example-test-tof-sensor/patch.xodp 	modified:   get-distance-mode/patch.cpp 	modified:   get-roi/patch.cpp 	modified:   init/patch.xodp 	modified:   project.xod 	modified:   set-distance-mode/patch.cpp 	modified:   set-roi/patch.xodp 	modified:   tof-sensor/patch.xodp 	modified:   vl53l1x-device/patch.cpp 	modified:   vl53l1x-device/patch.xodp

+++ get-roi/patch.cpp
@@ -7,7 +7,7 @@ node {
         // Get a pointer to the `VL53L1X` class instance
         auto sensor = getValue<input_DEV>(ctx);
 
-        uint8_t x, y;
+        uint16_t x, y;
         sensor->getROIxy(&x, &y);
         emitValue<output_X>(ctx, x);
         emitValue<output_Y>(ctx, y);
